Separate out SelfPlayerBoard into new component
